+++ packages/rhf-mui-components/src/mui-pickers/date-time/index.tsx
@@ -48,10 +48,10 @@ const RHFDateTimePicker = <T extends FieldValues>({
   formHelperTextProps,
   ...rest
 }: RHFDateTimePickerProps<T>) => {
-  const { dateAdapter, allLabelsAboveFormField } = useContext(RHFMuiConfigContext);
+  const { dateAdapter, allLabelsAboveFields } = useContext(RHFMuiConfigContext);
   const isLabelAboveFormField = keepLabelAboveFormField(
     showLabelAboveFormField,
-    allLabelsAboveFormField
+    allLabelsAboveFields
   );
   const fieldLabel = label ?? fieldNameToLabel(fieldName);
   const isError = Boolean(errorMessage);