+++ packages/rhf-mui-components/src/mui-pickers/date/index.tsx
@@ -57,10 +57,10 @@ const RHFDatePicker = <T extends FieldValues>({
   formHelperTextProps,
   ...rest
 }: RHFDatePickerProps<T>) => {
-  const { dateAdapter, allLabelsAboveFormField } = useContext(RHFMuiConfigContext);
+  const { dateAdapter, allLabelsAboveFields } = useContext(RHFMuiConfigContext);
   const isLabelAboveFormField = keepLabelAboveFormField(
     showLabelAboveFormField,
-    allLabelsAboveFormField
+    allLabelsAboveFields
   );
   const fieldLabel = label ?? fieldNameToLabel(fieldName);
   const isError = Boolean(errorMessage);